+++ CHANGELOG.md
@@ -83,6 +83,10 @@ and this project adheres to
   The main effect of this is, when post-styling a table with CSS, cell content
   remains visible. I.e. Z-ordering is more intuitive.
 
+### Fixed
+
+- The `ordering` graph attribute is now honored when set on subgraphs. #2503
+
 ## [15.1.0] – 2026-06-17
 
 ### Added